+++ src/Akkatecture/Aggregates/AggregateRootSettings.cs
@@ -21,6 +21,7 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System;
 using Akka.Configuration;
 using Akkatecture.Configuration;
 
@@ -31,6 +32,7 @@ namespace Akkatecture.Aggregates
         private static readonly string _section = "akkatecture.aggregate-root";
         public readonly bool UseDefaultEventRecover;
         public readonly bool UseDefaultSnapshotRecover;
+        public readonly TimeSpan SetReceiveTimeout;
 
         public AggregateRootSettings(Config config)
         {
@@ -39,6 +41,7 @@ namespace Akkatecture.Aggregates
 
             UseDefaultEventRecover = aggregateRootConfig.GetBoolean("use-default-event-recover");
             UseDefaultSnapshotRecover = aggregateRootConfig.GetBoolean("use-default-snapshot-recover");
+            SetReceiveTimeout = aggregateRootConfig.GetTimeSpan("set-receive-timeout");
         }
     }
 }