+++ src/Akkatecture/Jobs/Commands/ScheduleRepeatedly.cs
@@ -35,13 +35,12 @@ namespace Akkatecture.Jobs.Commands
 
         public ScheduleRepeatedly(
             TIdentity jobId,
-            ActorPath jobRunner,
             TJob job,
             TimeSpan interval,
             DateTime triggerDate,
             object ack = null,
             object nack = null)
-            : base(jobId, jobRunner, job, triggerDate, ack, nack)
+            : base(jobId, job, triggerDate, ack, nack)
         {
             if(interval == default) throw new ArgumentException(nameof(interval));
             
@@ -51,12 +50,12 @@ namespace Akkatecture.Jobs.Commands
         
         public override Schedule<TJob,TIdentity> WithAck(object ack)
         {
-            return new ScheduleRepeatedly<TJob, TIdentity>(JobId, JobRunner, Job, Interval, TriggerDate, ack, Nack);
+            return new ScheduleRepeatedly<TJob, TIdentity>(JobId, Job, Interval, TriggerDate, ack, Nack);
         }
         
         public override Schedule<TJob,TIdentity> WithNack(object nack)
         {
-            return new ScheduleRepeatedly<TJob, TIdentity>(JobId, JobRunner, Job, Interval, TriggerDate, Ack, nack);
+            return new ScheduleRepeatedly<TJob, TIdentity>(JobId, Job, Interval, TriggerDate, Ack, nack);
         }
     }
 }