+++ src/Akkatecture/Jobs/IJobManager.cs
@@ -1,0 +1,16 @@
+namespace Akkatecture.Jobs
+{
+    public interface IJobManager
+    {
+        
+    }
+
+    public interface IJobManager<TJob, TIdentity> : IJobManager
+        where TJob : IJob
+        where TIdentity : IJobId
+    {
+        
+    }
+
+
+}