+++ test/Akkatecture.Tests/UnitTests/Jobs/ScheduledJobsTests.cs
@@ -1,8 +1,10 @@
 using System;
 using System.ComponentModel;
+using System.Linq.Expressions;
 using Akka.Actor;
 using Akka.TestKit;
 using Akka.TestKit.Xunit2;
+using Akkatecture.Jobs;
 using Akkatecture.Jobs.Commands;
 using Akkatecture.TestHelpers.Jobs;
 using FluentAssertions;
@@ -26,34 +28,95 @@ namespace Akkatecture.Tests.UnitTests.Jobs
         [Category(Category)]
         public void SchedulingJob_For5minutes_DispatchesJobMessage()
         {
-            
-            var probe = CreateTestProbe("job-handler");
-            var jobId = TestJobId.New;
+            var probe = CreateTestProbe("job-probe");
             var scheduler = (TestScheduler) Sys.Scheduler;
-            var greeting = "Hi from the past";
-            var testJobScheduler = Sys.ActorOf(Props.Create(() => new TestJobScheduler()).WithDispatcher(CallingThreadDispatcher.Id), "test-job");
-            var when = DateTime.UtcNow.AddDays(1);
-            
+            var jobId = TestJobId.New;
+            var greeting = $"hi here here is a random guid {Guid.NewGuid()}";
             var job = new TestJob(greeting);
-            var schedule = new Schedule<TestJob, TestJobId>(jobId, probe.Ref.Path, job, when)
+            var when = DateTime.UtcNow.AddDays(1);
+            Expression<Func<TestJobScheduler>> testJobSchedulerExpression = () => new TestJobScheduler();
+            Expression<Func<TestJobRunner>> testJobRunnerExpression = () => new TestJobRunner(probe);
+            
+            var testJobManager = Sys.ActorOf(
+                Props.Create(() =>
+                new JobManager<TestJobScheduler, TestJobRunner, TestJob, TestJobId>(
+                    testJobSchedulerExpression,
+                    testJobRunnerExpression))
+                .WithDispatcher(CallingThreadDispatcher.Id));
+            
+            var schedule = new Schedule<TestJob, TestJobId>(jobId, job, when)
                 .WithAck(TestJobAck.Instance)
                 .WithNack(TestJobNack.Instance);
             
-            testJobScheduler.Tell(schedule, probe);
             
+            
+            testJobManager.Tell(schedule, probe);
             probe.ExpectMsg<TestJobAck>();
-            
             scheduler.AdvanceTo(when);
 
-            probe.ExpectMsg<TestJob>(x => x.Greeting == greeting);
+            
+            
+            probe.ExpectMsg<TestJobDone>(x =>
+            {
+                x.At.Should().BeCloseTo(when);
+                return x.Greeting == greeting;
+            });
+            scheduler.AdvanceTo(when.AddDays(1));
+            probe.ExpectNoMsg();
         }
         
         [Fact]
         [Category(Category)]
         public void SchedulingJob_ForEvery5minutes_DispatchesJobMessage()
         {
+            var probe = CreateTestProbe("job-probe");
+            var scheduler = (TestScheduler) Sys.Scheduler;
+            var jobId = TestJobId.New;
+            var greeting = $"hi here here is a random guid {Guid.NewGuid()}";
+            var job = new TestJob(greeting);
+            var when = DateTime.UtcNow.AddDays(1);
+            var fiveMinutes = TimeSpan.FromMinutes(5);
+            Expression<Func<TestJobScheduler>> testJobSchedulerExpression = () => new TestJobScheduler();
+            Expression<Func<TestJobRunner>> testJobRunnerExpression = () => new TestJobRunner(probe);
             
-            var probe = CreateTestProbe("job-handler");
+            var testJobManager = Sys.ActorOf(
+                Props.Create(() =>
+                        new JobManager<TestJobScheduler, TestJobRunner, TestJob, TestJobId>(
+                            testJobSchedulerExpression,
+                            testJobRunnerExpression))
+                    .WithDispatcher(CallingThreadDispatcher.Id));
+            
+            var schedule = new ScheduleRepeatedly<TestJob, TestJobId>(jobId, job, fiveMinutes, when)
+                .WithAck(TestJobAck.Instance)
+                .WithNack(TestJobNack.Instance);
+            
+            
+            
+            testJobManager.Tell(schedule, probe);
+            probe.ExpectMsg<TestJobAck>(TimeSpan.FromMinutes(1));
+            scheduler.AdvanceTo(when);
+            
+            
+            
+            probe.ExpectMsg<TestJobDone>(x =>
+            {
+                x.At.Should().BeCloseTo(when);
+                return x.Greeting == greeting;
+            }, TimeSpan.FromMinutes(1));
+            scheduler.Advance(fiveMinutes);
+            var t = this.Sys.Settings.System.Scheduler.Now.DateTime;
+            probe.ExpectMsg<TestJobDone>(x =>
+            {
+                x.At.Should().BeCloseTo(when.Add(fiveMinutes));
+                return x.Greeting == greeting;
+            }, TimeSpan.FromMinutes(1));
+            /*scheduler.Advance(fiveMinutes);
+            probe.ExpectMsg<TestJobDone>(x =>
+            {
+                x.At.Should().BeCloseTo(when);
+                return x.Greeting == greeting;
+            });*/
+            /*var probe = CreateTestProbe("job-handler");
             var jobRunner = Sys.ActorOf(Props.Create(() => new TestJobRunner(probe.Ref)));
             var jobId = TestJobId.New;
             var scheduler = (TestScheduler) Sys.Scheduler;
@@ -62,7 +125,7 @@ namespace Akkatecture.Tests.UnitTests.Jobs
             var when = DateTime.UtcNow.AddDays(1);
             
             var job = new TestJob(greeting);
-            var schedule = new ScheduleRepeatedly<TestJob, TestJobId>(jobId, jobRunner.Path, job, TimeSpan.FromMinutes(5), when)
+            var schedule = new ScheduleRepeatedly<TestJob, TestJobId>(jobId, job, TimeSpan.FromMinutes(5), when)
                 .WithAck(TestJobAck.Instance)
                 .WithNack(TestJobNack.Instance);
             
@@ -80,7 +143,7 @@ namespace Akkatecture.Tests.UnitTests.Jobs
             
             scheduler.Advance(TimeSpan.FromMinutes(5));    
 
-            probe.ExpectMsg<TestJobDone>(x => x.Greeting == greeting);
+            probe.ExpectMsg<TestJobDone>(x => x.Greeting == greeting);*/
         }
         
         [Fact]
@@ -98,7 +161,7 @@ namespace Akkatecture.Tests.UnitTests.Jobs
             var when = DateTime.UtcNow.AddMonths(1);
             
             var job = new TestJob(greeting);
-            var schedule = new ScheduleCron<TestJob, TestJobId>(jobId, jobRunner.Path, job, cronExpression, when)
+            var schedule = new ScheduleCron<TestJob, TestJobId>(jobId, job, cronExpression, when)
                 .WithAck(TestJobAck.Instance)
                 .WithNack(TestJobNack.Instance);
             
@@ -151,7 +214,7 @@ namespace Akkatecture.Tests.UnitTests.Jobs
             var when = DateTime.UtcNow.AddDays(1);
             
             var job = new TestJob(greeting);
-            var schedule = new ScheduleRepeatedly<TestJob, TestJobId>(jobId, jobRunner.Path, job, TimeSpan.FromMinutes(5), when)
+            var schedule = new ScheduleRepeatedly<TestJob, TestJobId>(jobId, job, TimeSpan.FromMinutes(5), when)
                 .WithAck(TestJobAck.Instance)
                 .WithNack(TestJobNack.Instance);